/backend --------------------- Now contains /requests folder with test HTTP request methods, useful for testing out the server's functionality directly.
Contains /build folder with "minified" version/production code of the
notes-app React application. The server uses Express middleware 'static'
that serves the React app when the server is visited at it's root
webpage.

/code
----------------
The services/notes.js no longer has a hard-coded base url. It is
understood that the frontend will be packaged for production and served
from the Express server, so its domain would be that of the server's.
Implemented updating importance of notes. Changes are synchronized
between frontend and backend.

Added setupProxy.js to ensure that development version of frontend can
still communicate with the Express server, as the source in
services/notes.js is such that only the production version of the
frontend can communicate with server. Proxy setup makes it possible for the frontend to work in dev
and prod modes.

B#	new file:   backend/build/static/js/main.60eab240.js.LICENSE.txt